Remove falsehood
I said "I though MacOS had inodes", implying that I thought MacOS did
not have inodes.  It turns out that it does, it's probably that I tried
installing inotify-tools on my Mac and when it said it couldn't be
installed (brew: Linux is required for this package), I convinced myself
that this was because MacOS didn't have inodes.